Reverting Monaco to 0.10.1 to avoid Mac issue where can't click anywhere except last line.
Also introduced a "monaco-test.html" that can easily be used for testing.

Simply create a snippet with:
```
window.top.location.href =
    "https://script-lab-react-beta.azurewebsites.net/monaco-test.html";
```
And then run it.  From there, you will be able to substitute any
version numbers/URLs that you want.

In the process of changing out the version numbers,
refactoring the logic a bit to make future adjustments easier
and more centralized.