Always include @types/custom-functions-runtime anytime you're in Excel
Today we already had this, but the functionality would only kick in AFTER you were in a custom functions solution (had `@customfunction` marking) AND had switched back and forth to the libraries.  Realistically, though, it does no harm to always be included whenever you're in Excel mode, and it helps with the non-golden-path.